Adapting the new package structure. (service layer config objects yet to be created)
