+++ components/org.wso2.carbon.identity.claim/src/main/java/org/wso2/carbon/identity/claim/exception/ClaimMappingReaderException.java
@@ -1,0 +1,39 @@
+/*
+*  Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+*
+*  WSO2 Inc. licenses this file to you under the Apache License,
+*  Version 2.0 (the "License"); you may not use this file except
+*  in compliance with the License.
+*  You may obtain a copy of the License at
+*
+*    http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing,
+* software distributed under the License is distributed on an
+* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+* KIND, either express or implied.  See the License for the
+* specific language governing permissions and limitations
+* under the License.
+*/
+
+package org.wso2.carbon.identity.claim.exception;
+
+/**
+ * CarbonSecurityDataHolderException class to handle
+ * NullPointerExceptions when retrieving data from
+ * IdentityMgtDataHolder
+ */
+public class ClaimMappingReaderException extends Exception {
+
+    public ClaimMappingReaderException() {
+        super();
+    }
+
+    public ClaimMappingReaderException(String message) {
+        super(message);
+    }
+
+    public ClaimMappingReaderException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
