+++ components/org.wso2.carbon.identity.claim/src/main/resources/api.identity.meta.claim.yaml
