--- components/org.wso2.carbon.identity.meta.claim.mgt/src/main/java/org/wso2/carbon/identity/meta/claim/mgt/mapping/profile/ClaimConfigEntry.java
@@ -1,128 +1,0 @@
-/*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wso2.carbon.identity.meta.claim.mgt.mapping.profile;
-
-import java.util.Collections;
-import java.util.Map;
-
-/**
- * Configuration of a profile
- */
-public class ClaimConfigEntry {
-    private String claimURI;
-    private String displayName;
-    private Boolean required;
-    private Boolean readonly;
-    private String verifier;
-    private String validator;
-    private String transformer;
-    private String defaultValue;
-    private String dataType;
-    private String regex;
-    private Map<String, String> properties;
-
-    public String getClaimURI() {
-        return claimURI;
-    }
-
-    public void setClaimURI(String claimURI) {
-        this.claimURI = claimURI;
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-
-    public void setDisplayName(String displayName) {
-        this.displayName = displayName;
-    }
-
-    public Boolean getRequired() {
-        return required;
-    }
-
-    public void setRequired(Boolean required) {
-        this.required = required;
-    }
-
-    public Boolean getReadonly() {
-        return readonly;
-    }
-
-    public void setReadonly(Boolean readonly) {
-        this.readonly = readonly;
-    }
-
-    public String getVerifier() {
-        return verifier;
-    }
-
-    public void setVerifier(String verifier) {
-        this.verifier = verifier;
-    }
-
-    public String getValidator() {
-        return validator;
-    }
-
-    public void setValidator(String validator) {
-        this.validator = validator;
-    }
-
-    public String getDefaultValue() {
-        return defaultValue;
-    }
-
-    public void setDefaultValue(String defaultValue) {
-        this.defaultValue = defaultValue;
-    }
-
-    public String getDataType() {
-        return dataType;
-    }
-
-    public void setDataType(String dataType) {
-        this.dataType = dataType;
-    }
-
-    public String getRegex() {
-        return regex;
-    }
-
-    public void setRegex(String regex) {
-        this.regex = regex;
-    }
-
-    public Map<String, String> getProperties() {
-        if (properties == null) {
-            return Collections.emptyMap();
-        } else {
-            return properties;
-        }
-    }
-
-    public void setProperties(Map<String, String> properties) {
-        this.properties = properties;
-    }
-
-    public String getTransformer() {
-        return transformer;
-    }
-
-    public void setTransformer(String transformer) {
-        this.transformer = transformer;
-    }
-}
