--- components/org.wso2.carbon.identity.meta.claim.mgt/src/main/java/org/wso2/carbon/identity/meta/claim/mgt/util/ProfileMgtConstants.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wso2.carbon.identity.meta.claim.mgt.util;
-
-/**
- *
- */
-public class ProfileMgtConstants {
-    public static final String PROFILE_MAPPING_FILE = "profile-mapping.yaml";
-    public static final String REQUIRED = "required";
-    public static final String READ_ONLY = "readonly";
-    public static final String VERIFIER = "verifier";
-    public static final String VALIDATOR = "validator";
-    public static final String DEFAULT_VALUE = "defaultValue";
-    public static final String DATA_TYPE = "dataType";
-    public static final String REGEX = "regex";
-
-}
