+++ feature/org.wso2.carbon.identity.claim.feature/resources/build.properties
