+++ feature/org.wso2.carbon.identity.claim.feature/resources/conf/claim-mapping.yaml
