+++ feature/org.wso2.carbon.identity.claim.feature/resources/conf/profile-mapping.yaml
