+++ feature/org.wso2.carbon.identity.claim.feature/resources/p2.inf
@@ -1,0 +1,2 @@
+instructions.configure = \
+org.eclipse.equinox.p2.touchpoint.natives.copy(source:${installFolder}/../../features/org.wso2.carbon.identity.claim_${feature.version}/conf,target:${installFolder}/../../../conf/identity,overwrite:true);\
