--- feature/org.wso2.carbon.identity.meta.claim.mgt.feature/pom.xml
@@ -1,142 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-  ~ Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-  ~
-  ~ WSO2 Inc. licenses this file to you under the Apache License,
-  ~ Version 2.0 (the "License"); you may not use this file except
-  ~ in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied. See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-    <parent>
-        <groupId>org.wso2.carbon.identity.mgt</groupId>
-        <artifactId>org.wso2.carbon.identity.mgt.parent</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>org.wso2.carbon.identity.meta.claim.mgt.feature</artifactId>
-
-    <packaging>carbon-feature</packaging>
-    <name>WSO2 Carbon Identity Management - Meta Claim Mgt - Feature</name>
-    <url>http://wso2.org</url>
-    <description>This feature contains the core bundles required by the Carbon Identity Meta Claim Management feature
-    </description>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.wso2.carbon.identity.mgt</groupId>
-            <artifactId>org.wso2.carbon.identity.meta.claim.mgt</artifactId>
-            <version>${carbon.identity.mgt.version}</version>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-            </resource>
-        </resources>
-        <plugins>
-            <plugin>
-                <artifactId>maven-resources-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>prefilter-resources</id>
-                        <phase>generate-resources</phase>
-                        <goals>
-                            <goal>copy-resources</goal>
-                        </goals>
-                        <configuration>
-                            <outputDirectory>src/main/resources</outputDirectory>
-                            <resources>
-                                <resource>
-                                    <directory>resources</directory>
-                                    <includes>
-                                        <include>conf/claim-mapping.yaml</include>
-                                        <include>conf/profile-mapping.yaml</include>
-                                        <include>p2.inf</include>
-                                        <include>build.properties</include>
-                                    </includes>
-                                </resource>
-                            </resources>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.wso2.carbon.maven</groupId>
-                <artifactId>carbon-feature-plugin</artifactId>
-                <version>${carbon.feature.plugin.version}</version>
-                <extensions>true</extensions>
-                <executions>
-                    <execution>
-                        <id>1-p2-feature-generation</id>
-                        <goals>
-                            <goal>generate</goal>
-                        </goals>
-                        <configuration>
-                            <propertyFile>../etc/feature.properties</propertyFile>
-                            <adviceFileContents>
-                                <advice>
-                                    <name>org.wso2.carbon.p2.category.type</name>
-                                    <value>server</value>
-                                </advice>
-                                <advice>
-                                    <name>org.eclipse.equinox.p2.type.group</name>
-                                    <value>false</value>
-                                </advice>
-                            </adviceFileContents>
-                            <bundles>
-                                <bundle>
-                                    <symbolicName>org.wso2.carbon.identity.meta.claim.mgt</symbolicName>
-                                    <version>${carbon.identity.mgt.version}</version>
-                                </bundle>
-                            </bundles>
-                            <importFeatures>
-                                <feature>
-                                    <id>org.wso2.carbon.identity.mgt.feature</id>
-                                    <version>${carbon.identity.mgt.version}</version>
-                                </feature>
-                            </importFeatures>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-antrun-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>clean_target</id>
-                        <phase>install</phase>
-                        <configuration>
-                            <target>
-                                <delete dir="src/main/resources"/>
-                                <delete dir="src/main"/>
-                                <delete dir="src"/>
-                            </target>
-                        </configuration>
-                        <goals>
-                            <goal>run</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
