--- feature/org.wso2.carbon.identity.meta.claim.mgt.feature/resources/p2.inf
@@ -1,2 +1,0 @@
-instructions.configure = \
-org.eclipse.equinox.p2.touchpoint.natives.copy(source:${installFolder}/../../features/org.wso2.carbon.identity.meta.claim.mgt_${feature.version}/conf,target:${installFolder}/../../../conf/identity,overwrite:true);\
