+++ pom.xml
@@ -157,7 +157,7 @@
             </dependency>
             <dependency>
                 <groupId>org.wso2.carbon.identity.mgt</groupId>
-                <artifactId>org.wso2.carbon.identity.meta.claim.mgt</artifactId>
+                <artifactId>org.wso2.carbon.identity.claim</artifactId>
                 <version>${carbon.identity.mgt.version}</version>
             </dependency>
             <dependency>
@@ -325,10 +325,10 @@
 
     <modules>
         <module>components/org.wso2.carbon.identity.mgt</module>
-        <module>components/org.wso2.carbon.identity.meta.claim.mgt</module>
+        <module>components/org.wso2.carbon.identity.claim</module>
         <module>components/org.wso2.carbon.identity.mgt.login.interceptor</module>
         <module>feature/org.wso2.carbon.identity.mgt.feature</module>
-        <module>feature/org.wso2.carbon.identity.meta.claim.mgt.feature</module>
+        <module>feature/org.wso2.carbon.identity.claim.feature</module>
         <module>feature/org.wso2.carbon.identity.mgt.login.feature</module>
         <module>tests</module>
     </modules>
