+++ tests/osgi-tests/pom.xml
@@ -78,7 +78,7 @@
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon.identity.mgt</groupId>
-            <artifactId>org.wso2.carbon.identity.meta.claim.mgt</artifactId>
+            <artifactId>org.wso2.carbon.identity.claim</artifactId>
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon.identity.mgt</groupId>
@@ -301,11 +301,11 @@
                                                     <fileset
                                                             dir="../../components/org.wso2.carbon.identity.mgt/target/classes"/>
                                                     <fileset
-                                                            dir="../../components/org.wso2.carbon.identity.meta.claim.mgt/target/classes"/>
+                                                            dir="../../components/org.wso2.carbon.identity.claim/target/classes"/>
                                                 </classfiles>
                                                 <sourcefiles encoding="UTF-8">
                                                     <fileset dir="../../components/org.wso2.carbon.identity.mgt/src"/>
-                                                    <fileset dir="../../components/org.wso2.carbon.identity.meta.claim.mgt/src"/>
+                                                    <fileset dir="../../components/org.wso2.carbon.identity.claim/src"/>
                                                 </sourcefiles>
                                             </structure>
                                             <html destdir="${project.build.directory}/site/jacoco"/>
