+++ tests/osgi-tests/src/test/java/org/wso2/carbon/identity/mgt/test/osgi/ClaimResolvingServiceTest.java
@@ -10,8 +10,8 @@ import org.osgi.framework.BundleContext;
 import org.testng.Assert;
 import org.testng.annotations.Listeners;
 import org.testng.annotations.Test;
-import org.wso2.carbon.identity.meta.claim.mgt.exception.ClaimResolvingServiceException;
-import org.wso2.carbon.identity.meta.claim.mgt.service.ClaimResolvingService;
+import org.wso2.carbon.identity.claim.ClaimResolvingService;
+import org.wso2.carbon.identity.claim.exception.ClaimResolvingServiceException;
 import org.wso2.carbon.identity.mgt.test.osgi.util.IdentityMgtOSGiTestUtils;
 import org.wso2.carbon.kernel.utils.CarbonServerInfo;
 
@@ -62,7 +62,6 @@ public class ClaimResolvingServiceTest {
                 "Claim mappings not read correctly");
 
     }
-
 
     @Test
     public void testDialectInheritance() throws ClaimResolvingServiceException {
