+++ tests/osgi-tests/src/test/java/org/wso2/carbon/identity/mgt/test/osgi/util/IdentityMgtOSGiTestUtils.java
@@ -87,7 +87,7 @@ public class IdentityMgtOSGiTestUtils {
                 .versionAsInProject());
         defaultOptionList.add(mavenBundle()
                 .groupId("org.wso2.carbon.identity.mgt")
-                .artifactId("org.wso2.carbon.identity.meta.claim.mgt")
+                .artifactId("org.wso2.carbon.identity.claim")
                 .versionAsInProject());
         defaultOptionList.add(mavenBundle()
                 .groupId("org.wso2.carbon.identity.mgt")
